% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick scatter plot in ggplot.}
\usage{
gbox(x, cat = NULL, vertical = TRUE)
}
\arguments{
\item{x}{This is the vector containing values to be plotted.}

\item{cat}{This is the vector containing the category for each value.
The default is NA: all of them belong to the same category.}

\item{vertical}{This is the orientation of the box plot. TRUE for vertical FALSE for horizontal.}

\item{point_size}{This is the size of the points that will be plotted.
The default is size TRUE: a vertical box plot.}
}
\value{
This function returns a ggplot box plot plot object.
}
\description{
This will graph a given vectors in a ggplot-style box plot with the
x-axis labeled "cat" if given a category vector and the y-axis labeled "x"
or vice versa if a horizontal box plot was selected.
}
\examples{
## Create a vertical box plot of x.
x <- rnorm(100)
gbox(x)

## Create a horizontal box plot of x by cat.
x <- rnorm(100)
cat <- sample(c("A","B"),100, replace = T)
gbox(x, cat, vertical = FALSE)

}
