% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{Create a quick histogram plot in ggplot.}
\usage{
ghist(x, border_color = "black", fill_color = "lightblue", bins = 10)
}
\arguments{
\item{x}{This is the vector to be plotted.}

\item{border_color}{This is the color of the bins' outline}

\item{fill_color}{This is the color of the bins}

\item{bins}{This is the number of bins for the histogram.
The default is 10 bins.}
}
\value{
This function returns a ggplot histogram object.
}
\description{
This will graph a given vector in a ggplot-style histogram plot with the
x-axis labeled "x" and the y-axis labeled "count".
}
\examples{
## Create a histogram plot of x.
x <- rnorm(100)
ghist(x)

}
